Fetching and mapping

+++ Chapter08/PostgreSQL/Maven/Basic/PojoTypes/src/main/java/com/classicmodels/pojo/RecordDepartment.java
@@ -1,3 +1,3 @@
 package com.classicmodels.pojo;
 
-public record RecordDepartment(String name, Short code, String[] topic) {}
+public record RecordDepartment(String name, Integer code, String[] topic) {}
